pyproject.toml
[tool.poetry]
package-mode = false
name = "backend"
version = "0.1.0"
description = ""
authors = ["maaz <maazoly1@gmail.com>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
django = "^5.1.1"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.9.0"
pytz = "^2024.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
pillow = "^10.4.0"
setuptools = "^75.1.0"
psycopg2 = "^2.9.9"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
pytest = "^8.3.3"
pytest-xdist = "^3.6.1"
pytest-django = "^4.9.0"
model-bakery = "^1.20.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
colorlog = "^6.8.2"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.api.settings"
django_find_project = true
python_files = "test_*.py"
testpaths = ["app"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
